format_subscription_list to broker/symbol/kucoin
