+++ src/broker/symbol/kucoin.rs
@@ -2,6 +2,7 @@ use crate::model::symbol::SymbolInfo;
 use crate::translator::traits::SymbolInfoTranslator;
 use kucoin_api::client::Kucoin;
 use kucoin_api::model::market::SymbolList;
+use kucoin_api::model::websocket::WSTopic;
 
 /// Uses the KuCoin API to generate a list of symbols
 pub async fn get_symbols(api: Kucoin) -> Vec<SymbolInfo> {
@@ -41,4 +42,47 @@ pub async fn get_symbols(api: Kucoin) -> Vec<SymbolInfo> {
         result.push(internal_symbol_info);
     }
     result
+}
+
+// TODO this bridges between API and the internal model, it should be placed in broker
+pub fn format_subscription_list(infos: &[SymbolInfo]) -> Vec<Vec<WSTopic>> {
+    // Extracts the symbol name from SynbolInfo
+    let symbols: Vec<String> = infos.iter().map(|info| info.symbol.clone()).collect();
+
+    // Sets up 2D array of max length 100
+    let max_sub_count = 100;
+    let mut hundred_arrays: Vec<Vec<String>> = Vec::new();
+    let mut hundred_array: Vec<String> = Vec::new();
+
+    // feed into the 2D array
+    for symbol in symbols {
+        hundred_array.push(symbol);
+        // 99 for the first one, because of the special BTC-USDT
+        if hundred_arrays.is_empty() && hundred_array.len() == max_sub_count - 1 {
+            hundred_arrays.push(hundred_array);
+            hundred_array = Vec::new();
+            continue;
+        }
+        // otherwise 100
+        if hundred_array.len() == max_sub_count {
+            hundred_arrays.push(hundred_array);
+            hundred_array = Vec::new();
+        }
+    }
+
+    // last array in current_subarray
+    if !hundred_array.is_empty() {
+        hundred_arrays.push(hundred_array);
+    }
+
+    let mut subs: Vec<Vec<WSTopic>> = Vec::new();
+    let mut sub: Vec<WSTopic> = Vec::new();
+    for sub_array in hundred_arrays {
+        sub.push(WSTopic::OrderBook(sub_array));
+        if sub.len() == 3 {
+            subs.push(sub);
+            sub = Vec::new();
+        }
+    }
+    subs
 }
