+++ src/broker/symbol/mod.rs
@@ -1,2 +1,4 @@
+/// Filter functions using internal models
 pub mod filter;
+/// Uses KuCoin API
 pub mod kucoin;
